Rename getChangesToRecentLessons file
